Put planner with threadpool/pulse dependencies behind feature
