+++ Cargo.toml
@@ -22,8 +22,12 @@ exclude = ["doc", ".travis.yml"]
 
 [dependencies]
 mopa = "0.2"
-pulse = "0.5"
-threadpool = "1.3"
+pulse = { version = "0.5", optional = true }
+threadpool = { version = "1.3", optional = true }
 fnv = "1.0"
 tuple_utils="0.2"
 atom = "0.3"
+
+[features]
+default = ["parallel"]
+parallel = ["threadpool", "pulse"]
